+++ src/main/java/cinema/tickets/booking/api/controller/SeatController.java
@@ -2,6 +2,7 @@ package cinema.tickets.booking.api.controller;
 
 import cinema.tickets.booking.api.dto.AudSeatReqDto;
 import cinema.tickets.booking.api.entity.Seat;
+import cinema.tickets.booking.api.exception.ResourceNotFoundException;
 import cinema.tickets.booking.api.service.SeatService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -24,7 +25,13 @@ public class SeatController {
 
     @GetMapping("/{id}")
     public ResponseEntity<Seat> getSeatById(@PathVariable int id) {
-        return new ResponseEntity<>(seatService.getById(id), HttpStatus.OK);
+        Seat res = seatService.getById(id);
+
+        if (res != null) {
+            return new ResponseEntity<>(seatService.getById(id), HttpStatus.OK);
+        } else {
+            throw new ResourceNotFoundException(String.format("Seat with id:%d not found", id));
+        }
     }
 
     @PostMapping("/")
@@ -34,13 +41,22 @@ public class SeatController {
 
     @PutMapping("/")
     public ResponseEntity<Seat> updateSeat(@RequestBody Seat seat) {
-        return new ResponseEntity<>(seatService.save(seat), HttpStatus.OK);
+        if (seatService.getById(seat.getId()) != null) {
+            return new ResponseEntity<>(seatService.save(seat), HttpStatus.OK);
+        } else {
+            throw new ResourceNotFoundException(String.format("Seat with id:%d not found", seat.getId()));
+        }
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<?> deleteSeat(@PathVariable int id) {
-        seatService.deleteById(id);
+        Seat res = seatService.getById(id);
 
-        return ResponseEntity.ok().build();
+        if (res != null) {
+            seatService.deleteById(id);
+            return ResponseEntity.ok().build();
+        } else {
+            throw new ResourceNotFoundException(String.format("Seat with id:%d not found", id));
+        }
     }
 }
